refactor : BaseEntity 위치 조정
- domain 하위로 위치 조정
