+++ backend/core/src/main/java/site/timecapsulearchive/core/domain/capsule/entity/Image.java
@@ -12,8 +12,7 @@ import jakarta.persistence.Table;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import site.timecapsulearchive.core.domain.capsule.entity.Capsule;
-import site.timecapsulearchive.core.global.entity.BaseEntity;
+import site.timecapsulearchive.core.domain.BaseEntity;
 
 @Entity
 @Getter
