+++ backend/core/src/main/java/site/timecapsulearchive/core/global/entity/NotificationCategory.java
@@ -11,6 +11,7 @@ import jakarta.persistence.ManyToOne;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
+import site.timecapsulearchive.core.domain.BaseEntity;
 
 @Entity
 @Getter
